bug:2224 change nb_image_page into smallint(3)
git-svn-id: http://piwigo.org/svn/trunk@10341 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ install/piwigo_structure-mysql.sql
@@ -404,7 +404,7 @@ CREATE TABLE `piwigo_user_group` (
 DROP TABLE IF EXISTS `piwigo_user_infos`;
 CREATE TABLE `piwigo_user_infos` (
   `user_id` smallint(5) NOT NULL default '0',
-  `nb_image_page` tinyint(3) unsigned NOT NULL default '15',
+  `nb_image_page` smallint(3) unsigned NOT NULL default '15',
   `status` enum('webmaster','admin','normal','generic','guest') NOT NULL default 'guest',
   `language` varchar(50) NOT NULL default 'en_UK',
   `maxwidth` smallint(6) default NULL,
